Revert "set the model to long"
This reverts commit 6bcdba333f1a84f4798c302a02a15cc190f47d09.

+++ app.py
@@ -67,7 +67,7 @@ def transcribe_audio():
         config = speech_v2.types.RecognitionConfig(
             auto_decoding_config=speech_v2.types.AutoDetectDecodingConfig(),
             language_codes=["si-LK"],
-            model="long",
+            model="short",
         )
         
         request_obj = speech_v2.types.RecognizeRequest(
